feat: update Bartholomew to be a Spin component
This commit updates Bartholomew to be a native Spin component, as opposed to
a Wagi module.

Specifically, the entry point of Bartholomew is now a typed function that takes
an HTTP request as argument, and returns an HTTP response:

```rust
pub fn render(req: Request) -> Result<Response> {...}
```

It also updates the static file server (referenced in the readme) to be the new
Spin static file server.

This commit temporarily removes the support for Brotli content encoding, which
will be re-enabled in a future commit.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>