feat(docs): add documentation website based on the Fermyon docs template
This commit adds a documentation website in the `docs/` directory for Bartholomew.
The website is based on the Fermyon docs website template, and is only focused on
the Bartholomew documentation, and not on showcasing all of Bartholomew's potential
features.
As such, not all pages and templates have been kept from the previous top-level
directories. Those should be part of the Bartholomew website template instead.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>